Added some basic unit tests for `getSignedUrlV1`
Added unit tests for `getSignedUrlV1` to verify
that the service account key used to sign the URL
is obtained from the appropriate source, SAM or Bond,
depending on the DRS URI value, if any.

Minor changes were required to `getSignedUrlV1.js`
to enable testing with sinon (AFAIK), and these changes
are consistent with similar code in the
`fileSummaryV1` and `martha_v2` endpoints.